AIIncidentIntelligencePlatform/sdk/requirements.txt
opentelemetry-api==1.22.0
opentelemetry-sdk==1.22.0
opentelemetry-exporter-otlp-proto-http==1.22.0
pydantic==2.5.3
